Admin panel added.

+++ Hastane Otomasyon/FrmPatientPanel.cs
@@ -196,7 +196,7 @@ namespace Hastane_Otomasyon
             }
             else
             {
-                MessageBox.Show("Please enter password and ID number !");
+                MessageBox.Show("Please fill password and ID number fields !");
 
             }
         }
@@ -340,19 +340,26 @@ namespace Hastane_Otomasyon
 
         private void BtnDelete_Click(object sender, EventArgs e)
         {
-            if (dataGridView1.SelectedRows.Count == 1)
+            DialogResult Secim = new DialogResult();
+
+            Secim = MessageBox.Show("Do you confirm the deletion?", "", MessageBoxButtons.YesNo, MessageBoxIcon.Question, MessageBoxDefaultButton.Button2);
+
+            if (Secim == DialogResult.Yes)
             {
-                con.Open();
-                SqlCommand cmd5 = new SqlCommand("delete from Appointment where AppointmentId = '" + dataGridView1.CurrentRow.Cells["AppointmentId"].Value.ToString() + "'", con);
-                cmd5.ExecuteNonQuery();
-                con.Close();
-                MessageBox.Show("Appointment Deleted !");
-                AppointmentCheck();
-                AppointmentList();
-            }
-            else
-            {
-                MessageBox.Show("Please select a valid appointment from the table below !");
+                if (dataGridView1.SelectedRows.Count == 1)
+                {
+                    con.Open();
+                    SqlCommand cmd5 = new SqlCommand("delete from Appointment where AppointmentId = '" + dataGridView1.CurrentRow.Cells["AppointmentId"].Value.ToString() + "'", con);
+                    cmd5.ExecuteNonQuery();
+                    con.Close();
+                    MessageBox.Show("Appointment Deleted !");
+                    AppointmentCheck();
+                    AppointmentList();
+                }
+                else
+                {
+                    MessageBox.Show("Please select a valid appointment from the table !");
+                }
             }
         }
     }
